Split up Client into component files
